# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Church Platform API — Phase 1 draft
  version: 0.1.0-draft
  description: Staff branch-read and branch account-request capture are implemented. Official membership operations remain draft and are not exposed.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /registrations/accounts:
    post:
      operationId: captureBranchAccountRegistration
      summary: Record a pending staff or member account request at an authorized South African branch
      description: Requires Member.Create for kind Member or Staff.Register for kind Staff at the stored branch. Browser clients first obtain a request token from /security/antiforgery and send it in X-CSRF-TOKEN. Does not create official membership, set credentials, activate an account or grant permissions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureAccountRegistrationRequest'
      responses:
        '201':
          description: Pending account request captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingAccountRegistration'
        '400':
          description: Invalid request or missing antiforgery token
        '401':
          description: Authentication required
        '403':
          description: Permission or branch scope denied
        '409':
          description: Account already exists for the email address
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: TheChg.Session
      description: First-party staff session; CSRF protection is required for browser writes.
...
  schemas:
    CaptureAccountRegistrationRequest:
      type: object
      additionalProperties: false
      required: [branchId, kind, email]
      properties:
        branchId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [Member, Staff]
        email:
          type: string
          format: email
          maxLength: 256
    PendingAccountRegistration:
      type: object
      additionalProperties: false
      required: [id, accountId, branchId, kind, status]
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        kind:
          type: string
          enum: [Member, Staff]
        status:
          type: string
          enum: [PendingReview]
...